chinese_ocr_db_crnn_mobile/requirements.txt
paddlepaddle
paddlehub
opencv-python-headless
Pillow